refactor: DebtMapper의 findByUserId 수정 #54
- DebtAccount에 id 필드 추가
- DebtMapper에서 findByUserId의 resultType을 entity로 수정
- DebtServiceImpl에서 Debt Entity를 DebtDTO로 변환하는 헬퍼 메소드 추가

+++ src/main/java/cheongsan/domain/debt/dto/DebtDTO.java
@@ -25,7 +25,7 @@ public class DebtDTO {
     private LocalDate loanStartDate;
     private LocalDate loanEndDate;
     private LocalDate nextPaymentDate;
-    private Integer gracePeriodMonths;
+    private Long gracePeriodMonths;
     private String repaymentMethod;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
